requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
dataclasses-json==0.6.1
Django==4.2.6
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.0
ibm-cos-sdk==2.13.2
ibm-cos-sdk-core==2.13.2
ibm-cos-sdk-s3transfer==2.13.2
ibm-watson-machine-learning==1.0.327
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.319
langsmith==0.0.49
loguru==0.7.2
lomond==0.3.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==1.5.3
pinecone-client==2.2.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.17.0
